requirements.txt
aiogram>=3.4
python-dotenv
apscheduler
